refactor: update notification type imports and restructure NotificationPopover path
